fix: split defaults to break client-side Prisma adapter import chain
setup-form-client.tsx (client component) imported constants from
models/defaults.ts which had a top-level import of lib/db.ts, pulling
@prisma/adapter-pg (with Node.js builtins) into the client bundle.

Split into defaults-data.ts (pure constants, no imports) and
defaults.ts (re-exports + server functions with prisma).
