+++ app/(auth)/self-hosted/setup-form-client.tsx
@@ -3,7 +3,7 @@ import { useState, useRef, useEffect, useCallback } from "react"
 import { FormSelectCurrency } from "@/components/forms/select-currency"
 import { FormInput } from "@/components/forms/simple"
 import { Button } from "@/components/ui/button"
-import { DEFAULT_CURRENCIES, DEFAULT_SETTINGS } from "@/models/defaults"
+import { DEFAULT_CURRENCIES, DEFAULT_SETTINGS } from "@/models/defaults-data"
 import { selfHostedGetStartedAction } from "../actions"
 import { FormSelect } from "@/components/forms/simple"
 import { PROVIDERS } from "@/lib/llm-providers"
